[RISCV][NFC] Factor out VLEN in the SiFive7 scheduling model (#143629)
In preparation of reusing SiFive7Model for sifive-x390, which has a VLEN
of 1024, it's better (and less chaotic) to factor out the VLEN parameter
from various of places first: the plan is to do a major overhaul on this
file in which all the `WriteRes` are going to be encapsulated in a big
`multiclass`, where VLEN is one of its template arguments. Such that we
can instantiate different scheduling models with different VLEN.

Before that happens, a placeholder defvar `SiFive7VLEN` is used instead
in this patch.

NFC.

Co-authored-by: Michael Maitland <michaeltmaitland@gmail.com>